vitres-pro/apps/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
supabase
weasyprint
apscheduler
resend
google-auth-oauthlib
google-api-python-client
python-multipart
pydantic-settings